test: add better file comparisons

+++ tests/util.sh
@@ -43,26 +43,32 @@ source ./tests/commands/upload_part.sh
 source ./tests/util_users.sh
 
 # recursively delete an AWS bucket
-# param:  bucket name
+# param:  client, bucket name
 # fail if error
 delete_bucket_recursive() {
   log 6 "delete_bucket_recursive"
-  assert [ $# -eq 2 ]
+  if [ $# -ne 2 ]; then
+    log 2 "'delete_bucket_recursive' requires client, bucket name"
+    return 1
+  fi
 
   local exit_code=0
   local error
   if [[ $1 == 's3' ]]; then
     error=$(aws --no-verify-ssl s3 rb s3://"$2" --force 2>&1) || exit_code="$?"
   elif [[ $1 == "aws" ]] || [[ $1 == 's3api' ]]; then
-    delete_bucket_recursive_s3api "$2"
+    if ! delete_bucket_recursive_s3api "$2"; then
+      log 2 "error deleting bucket recursively (s3api)"
+      return 1
+    fi
     return 0
   elif [[ $1 == "s3cmd" ]]; then
     error=$(s3cmd "${S3CMD_OPTS[@]}" --no-check-certificate rb s3://"$2" --recursive 2>&1) || exit_code="$?"
   elif [[ $1 == "mc" ]]; then
-    error=$(delete_bucket_recursive_mc "$2") || exit_code="$?"
+    error=$(delete_bucket_recursive_mc "$2" 2>&1) || exit_code="$?"
   else
-    log 2 "invalid command type '$1'"
-    assert [ 1 ]
+    log 2 "invalid client '$1'"
+    return 1
   fi
 
   if [ $exit_code -ne 0 ]; then
@@ -70,20 +76,25 @@ delete_bucket_recursive() {
       return 0
     else
       log 2 "error deleting bucket recursively: $error"
-      assert [ 1 ]
+      return 1
     fi
   fi
   return 0
 }
 
+# params:  bucket name
+# return 0 for success, 1 for error
 add_governance_bypass_policy() {
   if [[ $# -ne 1 ]]; then
-    log 2 "'add governance bypass policy' command requires command ID"
+    log 2 "'add governance bypass policy' command requires bucket name"
     return 1
   fi
   test_file_folder=$PWD
   if [[ -z "$GITHUB_ACTIONS" ]]; then
-    create_test_file_folder
+    if ! create_test_file_folder; then
+      log 2 "error creating test file folder"
+      return 1
+    fi
   fi
   cat <<EOF > "$test_file_folder/policy-bypass-governance.txt"
 {
@@ -102,10 +113,14 @@ EOF
     log 2 "error putting governance bypass policy"
     return 1
   fi
+  return 0
 }
 
 log_bucket_policy() {
-  assert [ $# -eq 1 ]
+  if [ $# -ne 1 ]; then
+    log 2 "'log_bucket_policy' requires bucket name"
+    return
+  fi
   if ! get_bucket_policy "s3api" "$1"; then
     log 2 "error getting bucket policy"
     return
@@ -183,28 +198,38 @@ check_and_disable_object_lock_config() {
 
 # restore bucket to pre-test state (or prep for deletion)
 # param: bucket name
-# fail on error
+# return 0 on success, 1 on error
 clear_bucket_s3api() {
   log 6 "clear_bucket_s3api"
-
-  assert [ $# -eq 1 ]
-
-  if [[ $LOG_LEVEL_INT -ge 5 ]]; then
-    run log_bucket_policy "$1"
-    assert_success "error logging bucket policy"
+  if [ $# -ne 1 ]; then
+    log 2 "'clear_bucket_s3api' requires bucket name"
+    return 1
   fi
 
-  run list_and_delete_objects "$1"
-  assert_success "error listing and delete objects"
+  if [[ $LOG_LEVEL_INT -ge 5 ]]; then
+    if ! log_bucket_policy "$1"; then
+      log 2 "error logging bucket policy"
+      return 1
+    fi
+  fi
 
-  run delete_bucket_policy "s3api" "$1"
-  assert_success "error deleting bucket policy"
+  if ! list_and_delete_objects "$1"; then
+    log 2 "error listing and deleting objects"
+    return 1
+  fi
+
+  if ! delete_bucket_policy "s3api" "$1"; then
+    log 2 "error deleting bucket policy"
+    return 1
+  fi
 
   #run check_ownership_rule_and_reset_acl "$1"
   #assert_success "error checking ownership rule and resetting acl"
 
-  run check_and_disable_object_lock_config "$1"
-  assert_success "error checking and disabling object lock config"
+  if ! check_and_disable_object_lock_config "$1"; then
+    log 2 "error checking and disabling object lock config"
+    return 1
+  fi
 
   #if ! change_bucket_owner "$AWS_ACCESS_KEY_ID" "$AWS_SECRET_ACCESS_KEY" "$1" "$AWS_ACCESS_KEY_ID"; then
   #  log 2 "error changing bucket owner back to root"
@@ -281,38 +306,51 @@ log_worm_protection() {
 }
 
 # params:  bucket name
-# fail if unable to delete bucket
+# return 0 if able to delete recursively, 1 if not
 delete_bucket_recursive_s3api() {
   log 6 "delete_bucket_recursive_s3api"
-  assert [ $# -eq 1 ]
+  if [ $# -ne 1 ]; then
+    log 2 "'delete_bucket_recursive_s3api' requires bucket name"
+    return 1
+  fi
 
-  clear_bucket_s3api "$1"
+  if ! clear_bucket_s3api "$1"; then
+    log 2 "error clearing bucket (s3api)"
+    return 1
+  fi
 
-  run delete_bucket 's3api' "$1"
-  assert_success "error deleting bucket"
-
+  if ! delete_bucket 's3api' "$1"; then
+    log 2 "error deleting bucket"
+    return 1
+  fi
   return 0
 }
 
 # params: client, bucket name
-# fail if error
+# return 0 on success, 1 on error
 delete_bucket_contents() {
   log 6 "delete_bucket_contents"
-  assert [ $# -eq 2 ]
+  if [ $# -ne 2 ]; then
+    log 2 "'delete_bucket_contents' requires client, bucket name"
+    return 1
+  fi
 
   local exit_code=0
   local error
   if [[ $1 == "aws" ]] || [[ $1 == 's3api' ]]; then
-    clear_bucket_s3api "$2"
-    return 0
+    if ! clear_bucket_s3api "$2"; then
+      log 2 "error clearing bucket (s3api)"
+      return 1
+    fi
   elif [[ $1 == "s3cmd" ]]; then
     delete_bucket_recursive "s3cmd" "$1"
-    return 0
   elif [[ $1 == "mc" ]]; then
     delete_bucket_recursive "mc" "$1"
-    return 0
+  else
+    log 2 "unrecognized client: '$1'"
+    return 1
   fi
-  assert [ 1 ]
+  return 0
 }
 
 # check if bucket exists
@@ -396,57 +434,73 @@ get_object_ownership_rule_and_update_acl() {
 }
 
 # params:  client, bucket name
-# fail if error
+# return 0 for success, 1 for error
 delete_bucket_or_contents() {
   log 6 "delete_bucket_or_contents"
-  assert [ $# -eq 2 ]
+  if [ $# -ne 2 ]; then
+    log 2 "'delete_bucket_or_contents' requires client, bucket name"
+    return 1
+  fi
   if [[ $RECREATE_BUCKETS == "false" ]]; then
-    delete_bucket_contents "$1" "$2"
+    if ! delete_bucket_contents "$1" "$2"; then
+      log 2 "error deleting bucket contents"
+      return 1
+    fi
 
-    run delete_bucket_policy "$1" "$2"
-    assert_success "error deleting bucket policies"
+    if ! delete_bucket_policy "$1" "$2"; then
+      log 2 "error deleting bucket policy"
+      return 1
+    fi
 
-    run get_object_ownership_rule_and_update_acl "$2"
-    assert_success "error getting object ownership rule and updating acl"
+    if ! get_object_ownership_rule_and_update_acl "$2"; then
+      log 2 "error getting object ownership rule and updating ACL"
+      return 1
+    fi
 
-    run abort_all_multipart_uploads "$2"
-    assert_success "error aborting multipart uploads"
-
+    if ! abort_all_multipart_uploads "$2"; then
+      log 2 "error aborting all multipart uploads"
+      return 1
+    fi
     log 5 "bucket contents, policy, ACL deletion success"
     return 0
   fi
-  run delete_bucket_recursive "$1" "$2"
-  assert_success "error with recursive bucket delete"
+  if ! delete_bucket_recursive "$1" "$2"; then
+    log 2 "error with recursive bucket delete"
+    return 1
+  fi
   log 5 "bucket deletion success"
   return 0
 }
 
 # params: client, bucket name
-# fail if unable to delete bucket (RECREATE_BUCKETS=true) or contents (RECREATE_BUCKETS=false)
+# return 0 for success, 1 for error
 delete_bucket_or_contents_if_exists() {
   log 6 "delete_bucket_or_contents_if_exists"
-
-  assert [ $# -eq 2 ]
+  if [ $# -ne 2 ]; then
+    log 2 "'delete_bucket_or_contents_if_exists' requires client, bucket name"
+    return 1
+  fi
 
   if bucket_exists "$1" "$2"; then
-    delete_bucket_or_contents "$1" "$2"
+    if ! delete_bucket_or_contents "$1" "$2"; then
+      log 2 "error deleting bucket and/or contents"
+      return 1
+    fi
     log 5 "bucket and/or bucket data deletion success"
     return 0
-  fi
-  if [[ $RECREATE_BUCKETS == "false" ]]; then
-    log 2 "When RECREATE_BUCKETS isn't set to \"true\", buckets should be pre-created by user"
-    assert [ 1 ]
   fi
   return 0
 }
 
+# params:  client, bucket name(s)
+# return 0 for success, 1 for failure
 setup_buckets() {
-  if [ $# -lt 1 ]; then
-    log 2 "'setup_buckets' command requires bucket names"
+  if [ $# -lt 2 ]; then
+    log 2 "'setup_buckets' command requires client, bucket names"
     return 1
   fi
-  for name in "$@"; do
-    if ! setup_bucket "$name"; then
+  for name in "${@:2}"; do
+    if ! setup_bucket "$1" "$name"; then
       log 2 "error setting up bucket $name"
       return 1
     fi
@@ -455,35 +509,40 @@ setup_buckets() {
 }
 
 # params:  client, bucket name
-# fail if bucket is not properly set up
+# return 0 on successful setup, 1 on error
 setup_bucket() {
   log 6 "setup_bucket"
-
-  assert [ $# -eq 2 ]
-
-  if [[ $1 == "s3cmd" ]]; then
-    log 5 "putting bucket ownership controls"
-    if bucket_exists "s3cmd" "$2"; then
-      run put_bucket_ownership_controls "$2" "BucketOwnerPreferred"
-      assert_success "error putting bucket ownership controls"
-    fi
+  if [ $# -ne 2 ]; then
+    log 2 "'setup_bucket' requires client, bucket name"
+    return 1
   fi
 
-  delete_bucket_or_contents_if_exists "$1" "$2"
+  if ! bucket_exists "$1" "$2" && [[ $RECREATE_BUCKETS == "false" ]]; then
+    log 2 "When RECREATE_BUCKETS isn't set to \"true\", buckets should be pre-created by user"
+    return 1
+  fi
+
+  if ! delete_bucket_or_contents_if_exists "$1" "$2"; then
+    log 2 "error deleting bucket or contents if they exist"
+    return 1
+  fi
 
   log 5 "util.setup_bucket: command type: $1, bucket name: $2"
   if [[ $RECREATE_BUCKETS == "true" ]]; then
-    run create_bucket "$1" "$2"
-    assert_success "error creating bucket"
-    log 5 "bucket creation success"
-
-    if [[ $1 == "s3cmd" ]]; then
-      log 5 "putting bucket ownership controls"
-      run put_bucket_ownership_controls "$2" "BucketOwnerPreferred"
-      assert_success "error putting bucket ownership controls"
+    if ! create_bucket "$1" "$2"; then
+      log 2 "error creating bucket"
+      return 1
     fi
   else
     log 5 "skipping bucket re-creation"
+  fi
+
+  if [[ $1 == "s3cmd" ]]; then
+    log 5 "putting bucket ownership controls"
+    if bucket_exists "s3cmd" "$2" && ! put_bucket_ownership_controls "$2" "BucketOwnerPreferred"; then
+      log 2 "error putting bucket ownership controls"
+      return 1
+    fi
   fi
   return 0
 }
